Remove /sdcard hack for shell users.
Using a /storage/sdcard0 symlink and a tmpfs staging area was used
to enable "/sdcard" paths for both zygote processes and the shell
user, but it breaks on devices with both emulated and physical
external storage.

Specifically, vold can't mount a physical sdcard onto the primary
storage symlink, and zygote can't re-bind-mount when secondary
storage is remounted after dropping privileges.

Bug: 7094858, 7064600
Change-Id: I988519af8024538d12ed2337b48d2df48dd2d2bc

Former-commit-id: e89411814298970c0fcdc034f403ed6127e43a3b